+++ src/test/java/io/beanmapper/testmodel/beansecuredfield/SFSourceAWithSecuredField.java
@@ -1,0 +1,10 @@
+package io.beanmapper.testmodel.beansecuredfield;
+
+import io.beanmapper.annotations.BeanSecuredProperty;
+
+public class SFSourceAWithSecuredField {
+
+    @BeanSecuredProperty("ADMIN")
+    public String name;
+
+}