+++ src/test/java/io/beanmapper/testmodel/beansecuredfield/SFTargetA.java
@@ -1,0 +1,7 @@
+package io.beanmapper.testmodel.beansecuredfield;
+
+public class SFTargetA {
+
+    public String name;
+
+}