+++ src/test/java/io/beanmapper/testmodel/beansecuredfield/SFTargetBWithSecuredField.java
@@ -1,0 +1,10 @@
+package io.beanmapper.testmodel.beansecuredfield;
+
+import io.beanmapper.annotations.BeanSecuredProperty;
+
+public class SFTargetBWithSecuredField {
+
+    @BeanSecuredProperty({"ADMIN", "DEV"})
+    public String name;
+
+}