+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Issue [#105](https://github.com/42BV/beanmapper/issues/105), **Ability to deal with @BeanSecuredProperty by delegating to a SecuredPropertyHandler**; when a field is tagged as @BeanSecuredProperty, BeanMapper will query its attached SecuredPropertyHandler. The handler will most likely be associated with a security implementation, such as Spring Security (not handled here). If no handler is present, access is granted by default.
 
 ## [2.3.2] - 2018-03-06
 ### Fixed